fixed wrong value for drug

+++ code/sql/create_drug.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS  Drug(
   ndc_no VARCHAR(20),
   combined_labeler_name VARCHAR(255),
-  dos_str FLOAT(9,2),
+  dos_str FLOAT(12,2),
   calc_base_wt_in_gm FLOAT,
   product_name VARCHAR(255),
   strength INTEGER,
